Exclude JCE signature files
The original Bouncycastle jars are signed with an JCE certificate to
enable registering them as a JCE provider. This prevents the
eclipse-jarsigner from signing these bundles.

We are waiting for fixes for
https://bugs.eclipse.org/bugs/show_bug.cgi?id=467064 requesting JCE
signing certificate to be ordered from Oracle
https://bugs.eclipse.org/bugs/show_bug.cgi?id=478088 requesting JCE
signing to be supported by signing service

Until these are available we should strip the original JCE signatures in
order to allow eclipse-jarsigner to sign these bundles.

Bug: 467064
Bug: 478088
Change-Id: I044331c6608605d48ab95d72bcb1a48e4c62ce30
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>