Bundle-ClassPath directory entries must contain the trailing slash.
In BundleComponent#createApiTypeContainer(String) when
ZipArchive.getEntry(..) reads in Bundle-ClassPath entries without a
trailing slash, they seem to be returned as ZipEntry objects whose names
also don't contain the trailing slash. As a result, isDirectory fails
(it checks for trailing slash in name) and a directory is wrongly
treated as a file.

Also make some small change in jffi so qualifiers match for both and
p2.inf continues to resolve the native bundle.

Change-Id: I732d5d7f27cb299bc3d5c762a77b06349a5fd9a8