Bug 511364: Add httpclient-win bundle and dependencies
This adds httpclient-win as org.apache.httpcomponents.httpclient.win.

The bundle does NOT export the org.apache.http.impl.client package,
since this conflicts with the same exported package from the
org.apache.httpcomponents.httpclient bundle, and given the limited
value of the provided content doesn't seem to justify the risk of
causing breakage with a split-package.

CQ: 12594
Change-Id: I0119b121162931ab9e073a9299a68963282fd780
Signed-off-by: Carsten Reckord <reckord@yatta.de>