Resolve missing files in com.github.jnr.jffi.
The instruction to include the p2.inf file was overriding the one from
ebr-maven-plugin for including content under 'resources'. Make it
possible for both of these to coexist.

Also make some small change in jffi.native so qualifiers match for both
and p2.inf continues to resolve the native bundle.

Change-Id: Icd4580359ca44c19ef16fb53fc5e7e1a69570978