Bug 501859: Exclude old R-build units from inclusion.
Some new bundles have dependencies to other bundles from the old
R-build. When the repository is generated, the category.xml ensures all
units get included even though they are not needed and cause
duplication.  'maven-artifactId' is a property name that all new
orbit-recipes bundles will have, but older content built through CVS
would not, so we use this to filter.

Change-Id: I4244a19675197203dd7cc53d974fa06275109d46