Bug 522740: Add Batik 1.9.1 into Orbit
This adds all the Batik 1.9.1 JARs into Orbit, with the exception of
the fop-transcoder-2.2.jar, which handles support for PDF output. That
one is not available from Maven Central and is handled separately, see
CQ #15532.

This commit also includes org.apache.xmlgraphics v2.2.0, which is not
technically part of Batik but is needed by it.

This packaging has several specific points which make it
"non-standard", as Batik 1.9 has several "split packages" issues which
require special handling.

First, the org.apache.batik.util package is split between the
org.apache.batik.util and org.apache.batik.constants bundles. This is
handled by:
- on the provider side, exporting the package with specific
  annotations, e.g.
  org.apache.batik.util;constants=split;mandatory:=constants
- on the client side, using explicit Require-Bundle for both
  bundles (note that this requires the
  "recipe.removeadditionalheaders" property to be set in the POMs, or
  BND will remove the Require-Bundle clauses from the final
  MANIFEST.MF).

Second, and more tricky, Batik uses types from the namespaces
org.w3c.dom and org.w3c.dom.events that it provides in
org.apache.batik.ext, except that these namespaces are already
exported by the JRE itself, with different contents. This causes
wiring issues, see the exchanges on the bugzilla starting from
http://eclip.se/522740#c37 for details.

The solution (more of a workaround) required the following packaging
tweaks:
- exporting the org.w3c.dom.events package from batik-ext with a
  specific annotation as above;
- disabling the generation of "uses" directives, as those would
  re-create implicit dependencies towards the non-annotated version of
  org.w3c.dom.events;
- embedding the org.w3c.dom.{ElementTraversal,Location,Window}
  directly in the JARs that need them by copying their sources in the
  appropriate src/main/java folders and changing the default BND
  export rules to keep them private to the JARs that need them.

Bug: 522740
Change-Id: Iff3337ad08256fc9cc30603e2166f328fd03cd0f
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>