Bug 526271 - logback.classic missing package import of logback.core
Some classes/packages of logback.core are not used in the code of the
logback.class bundle. Thus BND does not add these to the imported
packages. But in a logback configuration file, these classes might be
used (e.g. RollingFileAppender).

This Patch explicitly adds the missing packages to the bnd file.

Change-Id: I24b2fe97c36772d6ec3011885b1e8812397c3099
Signed-off-by: Christian Mohr <christian.mohr@cjt.de>