Add org.hamcrest.library 1.3.0
This bundle is already in Orbit built by the old build process.
Add this to orbit-recipes in order to rebuild it since signing
information in the old bundle isn't compatible with the signing of
org.hamcrest.core 1.3.0 which leads to runtime errors when using both
bundles on the same classpath. See bug 535043.

Bug: 535043
Change-Id: I064be690286218d576734bc8a01fbe9eaf267773
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>