Bug 549196 - Upgrade jackson stack to fix CVEs
Jackson 2.9.2 contains CVE 2019-12086.
Jackson 2.9.9.2 contains CVE 2019-14379 and
CVE 2019-14439.

To fix all this :
Update jackson-databind to 2.9.9.3 (represented as 2.9.93)
Update All other jackson modules to 2.9.9

Note :
There will be multiple warnings such as :

[WARNING] Manifest org.eclipse.orbit.bundles:com.fasterxml.jackson.core.
jackson-databind:eclipse-bundle-recipe:2.9.9.3 : Version for package com.
fasterxml.jackson.databind.jsonFormatVisitors is set to different values
in the source (2.9.9) and in the manifest (2.9.9.3). The version in the
manifest is not picked up by an other sibling bundles in this project or
projects that directly depend on this project

I think when jackson gets a version bump these should automatically go
away.

Change-Id: I89f376fd540bc4b57f785ca9cc35409de71ccb84
Signed-off-by: agola <anjali.gola@intel.com>