Bug 568379: com.sun.xml.bind 2.2.0
This bundle exists in CVS, but the uses directives are
missing, meaning it can resolve incorrectly. CQ 3843

Ideally this bundle would be removed, but until we
can get all consumers in simrel to update to newer
version we need to make this version work.

The issue is consumers who have this:

Require-Bundle: javax.xml.bind;bundle-version="2.2.0",
 com.sun.xml.bind;bundle-version="2.2.0"

Because there is no javax.xml.bind 2.3.3 that can satisfy the
above, having a 2.2.0 that wires correctly resolves
the problem.

Consumers who have:

Require-Bundle: com.sun.xml.bind;bundle-version="2.2.0"
Import-Package: javax.xml.bind;version="2.2.0"

will get either the 2.2.0 or the 2.3.3 versions of the
required bundles, but with the uses they will get
them consistently.


Change-Id: Idf06c5e187788d602fd74a2da762e3b3f46a0e6d