Add reload4j 1.2.19 as a successor of log4j 1.2.x
Initiated by Ceki Gülcü, the original author of Apache log4j 1.x, the
reload4j project [1] is a fork of Apache log4j version 1.2.17 with the
goal of fixing pressing security issues. It is intended as a drop-in
replacement for log4j version 1.2.17. By drop-in, we mean the
replacement of log4j.jar with reload4j.jar in your build with no source
code changes in .java files being necessary.

Use the same bundle name org.apache.log4j as the earlier log4j 1.2.15
available in Orbit to enable usage by Eclipse projects which currently
depend on the log4j 1.2.x which is EOL since years.

[1] https://reload4j.qos.ch/

Change-Id: I3991c7b72506b2e25349c965d0985ff017b3bfeb