Bug 569078: Make a limited version of bundles from Orbit for SimRel
Due to projects unintentionally depending on Orbit some bundles
(as listed in this pom.xml) are expected to be in SimRel even though
no project contributes them.

The intention is that each project that depends on these bundles
fix up their contribution, but in the meantime this workaround
is provided.

This is assigned against Bug 569078 because Bug 569078 is what
introduced orbit's direct contribution to SimRel as a workaround
that is now trying to be undone.

Change-Id: I800507bc9e88eb09d1d76c3e0be9b5e5ea149c22