Remove platform cyclic dependency
Some bundles require some OSGi bundles to be in the target
platform to resolve successful. Rather than depend on
all of eclipse platform which can have unintended consequences,
create a mirror of the parts of platform we are specifically
ok with having in the cycle in:

   releng/aggregation-mirror-osgi/pom.xml

and add that to the target platform for aggregating Orbit's
p2 update site.

An example of the unintended consequence can be seen in
https://git.eclipse.org/r/c/orbit/orbit-recipes/+/197009

Change-Id: Ieb91153f2a43e291136f765fd1c27d0b28332dcf