Bug 581148: Use clearly defined for all the bundles that use to use 22020
CQ 22020 was a placeholder that existed before Orbit supported
clearlydefined in the metadata. However now this bad reference
causes dash license tool to misreport.

All these bundles are actually Eclipse project (jakarta and glassfish),
including some which are the names from before Eclipse adopted them.

Many of them don't score 60+ on clearly defined, but that is ok because
they are Eclipse projects. See the bug where that is checked with
Wayne/Maria Teresa

Change-Id: I7f66c301979c9957cc9f5f105a7e47ee5017fc17